requirements.txt
fastapi==0.111.0
starlette==0.37.2
jinja2==3.1.4
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
streamlit==1.35.0
email-validator==2.1.1
